% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1gen.wrap.R
\name{formated_read_input}
\alias{formated_read_input}
\title{Function to read in the data}
\usage{
formated_read_input(data, response_type)
}
\arguments{
\item{data}{The data file we want to format. This is simply a data frame with
outcomes. We can input either a single outcome or multiple outcomes. The outcomes
each have a treatment and a result vector.}

\item{response_type}{The response type of the data. Can be a list of strings
if we have several outcomes to examine or a single string.}
}
\value{
The function returns the formated data.
}
\description{
A function that allows us to read in the data and fix any binomial
inputs to be 0 and 1.
}
