% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1gen.summary.R
\name{frequency_plot}
\alias{frequency_plot}
\title{Frequency plot for raw data}
\usage{
frequency_plot(nof1, xlab = "Outcomes", title = NULL, bins = 10)
}
\arguments{
\item{nof1}{nof1 object created using nof1.data}

\item{xlab}{x axis label}

\item{title}{title name}

\item{bins}{number of bins. To be used for continuous data}
}
\description{
Frequency plot for raw data
}
\examples{
Y <- laughter$Y
Treat <- laughter$Treat
nof1 <- nof1.data(Y, Treat, ncat = 11, baseline = 'Usual Routine', response = 'ordinal')
frequency_plot(nof1)
}
