% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1gen.summary.R
\name{make_raw_graphs}
\alias{make_raw_graphs}
\title{Raw Graphs and Tables}
\usage{
make_raw_graphs(graph, dataset, index, time = NULL, timestamp = NULL,
  xlab = NULL, title = NULL)
}
\arguments{
\item{graph}{(str) the specific table or graph to construct. Options include:
time_series_plot, frequency_plot, stacked_percent_barplot, and raw_table.}

\item{dataset}{the dataset used to create the table or graph}

\item{index}{(int) the outcome index in the dataset}

\item{...}{specific paramaters depending on table or graph to constuct}
}
\description{
\code{make_raw_graphs} allows the user to create a specific table or graph
for an outcome. Possible graphs include: time_series_plot, frequency_plot,
stacked_percent_barplot, and raw_table.
}
\examples{
make_raw_graphs("raw_table", diet_form, 1, title = "Daily Stool Consistency")
make_raw_graphs("stacked_percent_barplot", diet_form, 1, title = "Daily Stool Consistency")
}
