% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1gen.summary.R
\name{odds_ratio_plot}
\alias{odds_ratio_plot}
\title{Odds ratio plot}
\usage{
odds_ratio_plot(result.list, level = 0.95, title = "Odds Ratio Plot")
}
\arguments{
\item{result.list}{A list of objects with information about the simulation.
The objects are derived from the output of nof1.run}

\item{level}{The level defines the quantiles we want to examine.}

\item{title}{The title of the graph. Default is just to be: "Odds Ratio Plot"}
}
\description{
Creates a odds ratio plot for all the outcomes
}
