% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1gen.summary.R
\name{result_graphs}
\alias{result_graphs}
\title{Result Graphs}
\usage{
result_graphs(graph, model_result, multiple = TRUE,
  outcome_name = NULL, result.name = NULL, title = NULL, bins = 30,
  x_min = NULL, x_max = NULL, level = 0.95)
}
\arguments{
\item{graph}{A string which specifies the specific table or graph to construct.
Options include: kernel_density, odds_ratio_plot, probability_barplot.}

\item{model_result}{The analysis we want to make the graphs for. It is
the output of \code{wrap_single}}

\item{multiple}{A boolean that indicates if the \code{model_result} input has multiple
possible outcome. In other words, was \code{model_result} input created \code{\link{from wrap_all}}
or \code{\link{wrap_single}}? The default is FALSE meaning we only have one outcome in the
\code{model_result} input.}

\item{outcome_name}{A string that indicates the name of the outcome we wish to graph}

\item{...}{Specific paramaters depending on table or graph to constuct}
}
\description{
\code{result_graphs} allows the user to create a specific table or graph
for an outcome. Possible graphs include: kernel density plot, odds ratio plot,
and probability barplot.
}
