% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1gen.wrap.R
\name{summarize_all_nof1}
\alias{summarize_all_nof1}
\title{Function to present a summary of our results}
\usage{
summarize_all_nof1(data_result_list, treatment_names, alpha = 0.025)
}
\arguments{
\item{data_result_list}{A list of data files and result files. This is the
output of \code{wrap_all}}

\item{treatment_names}{The names of the different treatments in the trial.}

\item{alpha}{The alpha value for the confidence interval. If no value is
entered will give the 95\% confidence interval.}
}
\value{
The function returns some useful information about the simulation. It
returns the following for each outcome.
\item{input_mean}{The mean of the data inputed for each treatment type}
\item{input_median}{The median of the data inputed for each treatment type}
\item{treat_mean}{The mean of the data outputed for each treatment type}
\item{treat_median}{The median of the data outputed for each treatment type}
\item{coef_mean}{The mean of the data outputed for each coefficient}
\item{coef_median}{The median of the data outputed for each coefficient}
\item{treat_greater_zero}{The probability that the output value is greater
 than zero for each treatment type}
\item{coef_greater_zero}{The probability that the output value is greater
 than zero for each coefficient}
\item{treat_confidence}{The confidence interval of the input data
set of the output value for each treatment type}
\item{coef_confidence}{The confidence interval of the input data set
of the output value for each coefficient}
}
\description{
A neat function to summarize the results when we run a simulation for each outcome
}
