% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1gen.wrap.R
\name{washout}
\alias{washout}
\title{Washout function}
\usage{
washout(read_data, metadata)
}
\arguments{
\item{read_data}{The data we want to apply a washout out}

\item{metadata}{The metadata section of a file. This is important as it contains
information about the length of the trial.}
}
\value{
The function returns the data with a washout period implemented.
}
\description{
Allows the user to implement a washout period in the dataset. To do this,
the function calculates the length of the trial and then will replace some of
the original values with NULL with a different frequency depending on how often
the data was collected.
}
