% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1gen.wrap.R
\name{wrap_all}
\alias{wrap_all}
\title{Function to run the analysis on all outcomes}
\usage{
wrap_all(dataset, response_list, washout = TRUE)
}
\arguments{
\item{dataset}{The dataset the analysis will be conducted on.}

\item{result_list}{A list of all the models to run. Options are: normal,
binomial, poisson, or ordinal.}
}
\value{
The function returns some useful information about the simulation
 as well as information about the simulation.
\item{system_info}{Provides information about the clincical trial conducted}
\item{user_id}{The user id for the particular patient whose data was analyzed}
\item{trigger}{The trigger the study was examining}
\item{design}{How the study was designed. How many weeks of treatment A? Of
 treatment B?}
\item{nof_treat}{The number of different treatments that were admininstered}
\item{data}{This the the data file that was constructed using \code{nof1.data}}
\item{samples}{This is the result of \code{nof1.run}. It is the result of the simulation}
}
\description{
\code{wrap_all} allows the user to run an anaylsis on a specific data set.
}
