% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1gen.wrap.R
\name{wrap_helper}
\alias{wrap_helper}
\title{Helper function used to run the models}
\usage{
wrap_helper(specific_data, response_type)
}
\arguments{
\item{specific_data}{The specific part of the dataset the analysis will be conducted on.}

\item{response_type}{The type of model we want to simulate. Can be normal,
binomial, poisson, or ordinal.}
}
\value{
The function returns a list of two objects.
\item{nof1_out}{This the the data file that was constructed using \code{nof1.data}}
\item{result_out}{This is the result of \code{nof1.run}. It is the result of the simulations}
}
\description{
Allows the user to run an anaylsis on a specific data set.
}
