% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1gen.wrap.R
\name{gen_wrap}
\alias{gen_wrap}
\title{Function to run the analysis}
\usage{
gen_wrap(dataset)
}
\arguments{
\item{dataset}{the dataset the analysis will be conducted on.}
}
\value{
The function returns some useful information about the simulation.
}
\description{
\code{gen_wrap} allows the user to run an anaylsis on a specific data set.
}
\examples{
# We can run the simulataions and get the result using gen_wrap.
result_afib <- gen_wrap(afib_form)
result_diet <- gen_wrap(diet_form)

# We can then also take the results and convert them into a .json format.
# The same format as our input files were.
output_afib <- toJSON(result_afib, pretty = TRUE, UTC = TRUE, auto_unbox = TRUE, na = NULL)
output_diet <- toJSON(result_diet, pretty = TRUE, UTC = TRUE, auto_unbox = TRUE, na = NULL)
}
