% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1gen.wrap.R
\name{washout}
\alias{washout}
\title{Washout function}
\usage{
washout(read_data, num_outcomes, set_to_null = NULL)
}
\arguments{
\item{read_data}{The data we want to apply a washout out}

\item{num_outcomes}{An integer, representative of the number of observations
in the read_data input.}

\item{set_to_null}{A string (or list of strings), which indicates how many
observations will be set to NA if a washout period is implemented. The default
is set to NULL.}
}
\value{
The function returns the data with a washout period implemented.
}
\description{
Allows the user to implement a washout period in the dataset. To do this,
the function calculates the length of the trial and then will replace some of
the original values with NULL with a different frequency depending on how often
the data was collected.
}
