% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1gen.wrap.R
\name{wrap_single}
\alias{wrap_single}
\title{Function to run the analysis on a single outcome}
\usage{
wrap_single(dataset, outcome_name, response_type, washout = TRUE,
  set_to_null = NULL)
}
\arguments{
\item{dataset}{The dataset the analysis will be conducted on.}

\item{outcome_name}{The name of the outcome the analysis will be run on.}

\item{response_type}{The type of model to run. Options are: normal,
binomial, poisson, or ordinal.}

\item{washout}{A boolean, which indicates whether a washout period will
be used or not. The default is set to TRUE.}

\item{set_to_null}{A string, which indicates how many observations will
be set to NA if a washout period is implemented. The default is set
to NULL.}
}
\value{
The function returns some useful information about the simulation
 as well as information about the simulation itself.
\item{system_info}{Provides information about the clincical trial conducted}
\item{user_id}{The user id for the particular patient whose data was analyzed}
\item{treatments}{The names of the treatments in the trial}
\item{nof_treat}{The number of different treatments that were admininstered}
\item{design}{How the study was designed. How many weeks of treatment A? Of
 treatment B?}
\item{model_results}{This is a list which contains the data file that was
constructed using \code{nof1.data} and the result file which was created
with \code{nof1.run}}
}
\description{
\code{wrap_single} allows the user to run an anaylsis
 on a specific outcome in a specific data set.
}
